Create custom class for dynamic background
