+++ App/connect_four/lib/Screens/Login/login_screen.dart
@@ -1,40 +1,6 @@
-import 'package:connect_four/background/dynamic_background.dart';
-import 'package:flutter/material.dart';
+import 'package:connect_four/Screens/Login/login_body.dart';
+import 'package:connect_four/background/background_widget.dart';
 
-import 'body.dart';
-
-class LoginScreen extends StatefulWidget {
-  @override
-  _LoginScreenState createState() => _LoginScreenState();
-}
-
-class _LoginScreenState extends State<LoginScreen>
-    with SingleTickerProviderStateMixin {
-  AnimationController animationController;
-
-  @override
-  void initState() {
-    animationController =
-        AnimationController(vsync: this, duration: Duration(seconds: 5));
-    super.initState();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    Future.delayed(Duration(milliseconds: 1000), () {
-      animationController.forward();
-    });
-    return Scaffold(
-      body: Stack(
-        children: [
-          SizedBox.expand(
-            child: CustomPaint(
-              painter: BackgroundPainter(animation: animationController.view),
-            ),
-          ),
-          Body()
-        ],
-      ),
-    );
-  }
+class LoginScreen extends MyDynamicStateful {
+  LoginScreen() : super(statelessWidget: LoginBody(), moveFromUpToDown: true);
 }
