--- App/connect_four/lib/Screens/Signup/body.dart
@@ -1,50 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:connect_four/Screens/Login/login_screen.dart';
-import 'package:connect_four/components/already_have_an_account_acheck.dart';
-import 'package:connect_four/components/rounded_button.dart';
-import 'package:connect_four/components/rounded_input_field.dart';
-import 'package:connect_four/components/rounded_password_field.dart';
-
-class Body extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    Size size = MediaQuery.of(context).size;
-    return SingleChildScrollView(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Text(
-              "SIGNUP",
-              style: TextStyle(fontWeight: FontWeight.bold),
-            ),
-            SizedBox(height: size.height * 0.03),
-            RoundedInputField(
-              hintText: "Your Email",
-              onChanged: (value) {},
-            ),
-            RoundedPasswordField(
-              onChanged: (value) {},
-            ),
-            RoundedButton(
-              text: "SIGNUP",
-              press: () {},
-            ),
-            SizedBox(height: size.height * 0.03),
-            AlreadyHaveAnAccountCheck(
-              login: false,
-              press: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                    builder: (context) {
-                      return LoginScreen();
-                    },
-                  ),
-                );
-              },
-            ),
-          ],
-        ),
-    );
-  }
-}
