+++ App/connect_four/lib/Screens/Welcome/welcome_screen.dart
@@ -1,7 +1,7 @@
 import 'package:connect_four/background/background_widget.dart';
 
-import 'body.dart';
+import 'welcome_body.dart';
 
 class WelcomeScreen extends MyDynamicStateful {
-  WelcomeScreen() : super(statelessWidget: Body(), startState: false);
+  WelcomeScreen() : super(statelessWidget: WelcomeBody(), moveFromUpToDown: false);
 }
